// src/app/models/images.model.ts
export class Images {
    name:                    string  =  "";
    namespace:               string  =  "";
    creationTimestamp:         Date  =  new Date;
    labels:                      {}  =  {};
    type:                    string  =  "";
    readableName:            string  =  "";
    readableDescription:     string  =  "";
    value:                   string  =  "";
    credentials:             string  =  "";
}